RavenDB-13106 address PR comments
