+++ src/Raven.Server/Documents/Replication/Outgoing/OutgoingMigrationReplicationHandler.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.IO;
-using Raven.Client.Documents.Operations.Replication;
-using Raven.Client.Documents.Replication;
-using Raven.Client.ServerWide;
+﻿using System.IO;
 using Raven.Client.ServerWide.Commands;
 using Raven.Server.Documents.Replication.Senders;
-using Raven.Server.Utils;
 using Sparrow.Logging;
 
 namespace Raven.Server.Documents.Replication.Outgoing
