+++ src/Raven.Server/ServerWide/Maintenance/Sharding/BucketsMigrator.cs
@@ -2,12 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using Raven.Client.ServerWide;
-using Raven.Client.Util;
-using Raven.Server.ServerWide.Commands;
 using Raven.Server.Utils;
-using Sparrow;
 
-namespace Raven.Server.ServerWide.Maintenance
+namespace Raven.Server.ServerWide.Maintenance.Sharding
 {
     public static class BucketsMigrator
     {
@@ -186,56 +183,6 @@ namespace Raven.Server.ServerWide.Maintenance
                     }
                 }
             }
-        }
-    }
-
-    public class MergedDatabaseStatusReport
-    {
-        public Dictionary<ShardNumber, DatabaseStatusReport> MergedReport = new Dictionary<ShardNumber, DatabaseStatusReport>();
-
-        public ShardReport[] GetShardsReports => MergedReport.Select(r => new ShardReport
-        {
-            Shard = r.Key,
-            ReportPerBucket = r.Value.ReportPerBucket
-        }).ToArray();
-    }
-
-    public class ShardMigrationResult
-    {
-        public string Database;
-        public string RaftId = RaftIdGenerator.NewId();
-
-        public int SourceShard;
-        public int DestinationShard;
-        public BucketNumber Bucket;
-
-        public StartBucketMigrationCommand GetMigrationCommand => new StartBucketMigrationCommand(Database, RaftId)
-        {
-            Bucket = Bucket,
-            SourceShard = SourceShard,
-            DestinationShard = DestinationShard,
-        };
-
-        public override string ToString()
-        {
-            return $"Migrate bucket '{Bucket}' from '{SourceShard}' to '{DestinationShard}'";
-        }
-
-        public override int GetHashCode() => (int)Hashing.XXHash64.Calculate(new int[] { SourceShard, DestinationShard, Bucket });
-
-        public override bool Equals(object? obj)
-        {
-            if (obj == null)
-                return false;
-            
-            if (obj is ShardMigrationResult result == false)
-                return false;
-
-
-            return result.SourceShard == SourceShard && 
-                   result.DestinationShard == DestinationShard && 
-                   result.Bucket == Bucket;
-
         }
     }
