+++ src/Raven.Server/ServerWide/Maintenance/Sharding/MergedDatabaseStatusReport.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace Raven.Server.ServerWide.Maintenance.Sharding
+{
+    public class MergedDatabaseStatusReport
+    {
+        public Dictionary<ShardNumber, DatabaseStatusReport> MergedReport = new Dictionary<ShardNumber, DatabaseStatusReport>();
+
+        public ShardReport[] GetShardsReports => MergedReport.Select(r => new ShardReport
+        {
+            Shard = r.Key,
+            ReportPerBucket = r.Value.ReportPerBucket
+        }).ToArray();
+    }
+}
